adjusted expected.txt and adjusted to fit checkstyle
