+++ src/main/java/seedu/duke/DayMap.java
@@ -37,11 +37,6 @@ public class DayMap {
         alist = this.getActivityList(dateTime);
 
         alist.addActivity(activity);
-//        if (activityType.equals("food")) {
-//            alist.addFood(activityDescription, calories);
-//        } else if (activityType.equals("exercise")) {
-//            alist.addExercise(activityDescription, calories);
-//        }
     }
 
     public ArrayList getArrayList(LocalDateTime dateTime) {
@@ -63,7 +58,7 @@ public class DayMap {
         }
     }
 
-    public HashMap<LocalDate, ActivityList> getHashMap(){
+    public HashMap<LocalDate, ActivityList> getHashMap() {
         return dayMap;
     }
