+++ src/main/java/seedu/duke/Parser.java
@@ -1,9 +1,9 @@
 package seedu.duke;
 
-import seedu.duke.Commands.AddExerciseCommand;
-import seedu.duke.Commands.AddFoodCommand;
-import seedu.duke.Commands.Command;
-import seedu.duke.Commands.ListCommand;
+import seedu.duke.command.AddExerciseCommand;
+import seedu.duke.command.AddFoodCommand;
+import seedu.duke.command.Command;
+import seedu.duke.command.ListCommand;
 import seedu.duke.userprofile.Initialiseuser;
 import seedu.duke.userprofile.Userinfo;
 
@@ -98,19 +98,19 @@ public class Parser {
                 //System.out.println("Current number of activities is: " + calList.getSizeOfActivityList(date));
                 // calList.getSizeOfActivityList(date));
                 return new AddFoodCommand(arguments[1], calories);
-            }
-            else if (arguments[1].startsWith("e/")) {
+
+            } else if (arguments[1].startsWith("e/")) {
                 int calorieIndex = arguments[1].indexOf("c/");
                 int calories = Integer.parseInt(arguments[1].substring(calorieIndex + 2).trim());
                 arguments[1] = arguments[1].substring(2, calorieIndex - 1).trim();
-//            calList.addActivity(date, arguments[1], calories, "exercise"); //daymap equivalent
+                //calList.addActivity(date, arguments[1], calories, "exercise"); //daymap equivalent
                 return new AddExerciseCommand(arguments[1], calories);
                 //System.out.println("Current number of activities is: " + calList.getSizeOfActivityList(date));
             }
-        }catch (NullPointerException | StringIndexOutOfBoundsException e) {
+        } catch (NullPointerException | StringIndexOutOfBoundsException e) {
             return null;
         }
-    return null;
+        return null;
     }
