+++ src/main/java/seedu/duke/command/Command.java
@@ -1,4 +1,4 @@
-package seedu.duke.Commands;
+package seedu.duke.command;
 
 import seedu.duke.DayMap;
